Significant changes to style + Bootstrap 3.0
This change makes significant changes to the way the parties page works.
It now looks and works completely differently.
It also does a lot more on the database server side (though with only 3 users
this shouldn't be a big problem!)

Tests
------------
Test that the Parties page appears to work ok in Firefox, Chrome, and iPad
Though it is a bit slow on the iPad!