Adding contacts for parties
This changeset adds support for contacts to parties.
You can add new contacts and move NPCs between the contacts and the party.

You can't currently delete from the contacts.

Tests
--------------
Test that the contacts stuff seems to work ok
Test that you can go to the char page for a new contact.
Test that you can add a new contact to the party and the party still works ok.