Fix text not hiding on tabs on phone
Bootstrap 3.0 changed the hidden-phone to hidden-xs.
Worse than that, hidden-xs only works on block-level items, so doesn't work for inline stuff like span.
So this changeset includes a modification to ui-bootstrap for the new text, plus some css from github that provides hidden-inline-xs

Tests
------------
Test that the text now disappears on the phone (leaving only the icons)