Fix bug with ml-relative not showing the correct content
The ml-relative directive was not forcing the input to be numbers, so the comparisons
were sometimes happening as strings and sometimes as numbers. This leads to weird
issues where e.g. 12 current -v- 5 base was showing red because "12" is leagraphically
less than "5".

Also improved the checks for using current, as previously the current as number 0 would
actually be ignored in favour of the base, when actually it should be included.

Tests
------------
Test that numbers now show the correct value and in the correct colour.