Fix problem where changes didn't persist on party screen until reload
The code is kinda badly designed at the moment so that the character controller
has lots of items at scope level rather than having them inside the char object.

So this means when we edit things in the character, although they are saved to
the database, when we go away and come back the base character is the old
data and we appear to lose the changes (though a data refresh would fix it).

This fixes that by reloading the old data. Which is a worse fix than just amalgamating
everything under the char object and not keeping copies of everything around etc.
But for now it will do.

Tests
------------
Test that changes to wounds and effects are now visibly persistent