Fix previous fix of single-shot weapon drain
The pervious fix didn't use enough ||'s so instead of doing a logical or (drain OR 1) it was doing a
bitwise or and basically adding 1 to everything.

This fixes it by using both || to do a logical OR as was intended.

Tests
------------
Test that ammo consumption is now correct for single-shot
weapons without drain, and normal weapons with drain