Introducing the ability to give items to other characters in the party
This currently allows giving weapons and armour.
AoE weapons are slightly special because you give one of your pile rather than the whole
thing.

Tests
------------
This works ok in Chrome, and Safari in iOS7
It doesn't work in Firefox or Safari in iOS 5.1.1 - something to do with the select probably.