Reworking "give" template to use buttons
As the select control wasn't working in iOS5, this reworks the control to use
buttons so that you can just click on the button of where you want it to go.

This does still have a bug that it sometimes re-loads to the main page of the
app in iOS5, but that's a more minor problem

Tests
-------------
Test that giving now works ok in firefox and iOS5.