Add realtime updates to the app
This relies on two parts:
1) dpd.on(...) to do the changes when reported
2) emit('char:changed',...) everywhere in the database when items are added, modified, or deleted.

There is also some extra complication in that Heroku kills sockets after 60s so this also adds some
keepalive code to keep sockets open. Its designed so that the client app makes a request, and
the server sends a keepalive to everyone over the socket. On top of this, its designed to not make
too many requests, so it doesn't make a keepalive if it received one from someone else recently.
Finally, there is some code to randomise the timeouts for the keepalive requests because if they
are too similar in time then all the requests get made before the first socket callback is received.

Tests
------------------
Test that modifying items causes another client to automatically update. Tested with:
* Stats, wounds, effects
* ammo usage
* giving weapons to others
Confirm that these effect both the data for the selected char (if that char), and the
summary in the selection list (if affects health).